♻️ refactor ApolloWrapper module
