+++ frontend/src/providers/ApolloWrapper/links/http.ts
@@ -1,0 +1,8 @@
+import { createHttpLink } from "@apollo/client";
+import config from "src/config";
+
+export default function useHttpLink() {
+  return createHttpLink({
+    uri: `${config.HASURA_BASE_URL}/v1/graphql`,
+  });
+}
