+++ frontend/src/providers/ApolloWrapper/links/retry.ts
@@ -1,0 +1,5 @@
+import { RetryLink } from "@apollo/client/link/retry";
+
+export default function useRetryLink() {
+  return new RetryLink();
+}
